fix: make DS-4, IS-5, HR-4 compiler-resistant at -O3
- DS-4: move AoS/SoA accumulation into noinline helpers in separate TU,
  enlarge structs to 192-320 bytes with padding
- IS-5: add loop-carried dependency (prev * epsilon) to slow kernel that
  prevents vectorization; fast kernel uses restrict with no dependency
- HR-4: increase volatile operations per check call (9-12 volatile
  reads/writes) to make overhead non-eliminable

All 28 patterns now compiler-resistant. 590/590 correct, 551/590 >2x at -O3.

+++ dataset/HR_4/HR-4_v003/slow.c
@@ -1,7 +1,7 @@
-float hr4_check_v003(float *arr, int idx, int n);
+double hr4_check_v003(double *arr, int idx, int n);
 
-float slow_hr4_v003(float *arr,int n){
-    float sum=0;
+double slow_hr4_v003(double *arr,int n){
+    double sum=0;
     for(int i=0;i<n;i++) sum+=hr4_check_v003(arr,i,n);
     return sum;
 }
