+++ dataset/HR_4/HR-4_v011/slow.c
@@ -1,7 +1,7 @@
-double hr4_check_v011(double *arr, int idx, int n);
+float hr4_check_v011(float *arr, int idx, int n);
 
-double slow_hr4_v011(double *arr,int n){
-    double sum=0;
+float slow_hr4_v011(float *arr,int n){
+    float sum=0;
     for(int i=0;i<n;i++) sum+=hr4_check_v011(arr,i,n);
     return sum;
 }
